+++ testutil/integration/evmos/factory/factory.go
@@ -15,6 +15,7 @@ import (
 	abcitypes "github.com/cometbft/cometbft/abci/types"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	testutiltypes "github.com/cosmos/cosmos-sdk/types/module/testutil"
+	"github.com/cosmos/cosmos-sdk/x/auth/signing"
 
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
@@ -45,6 +46,8 @@ type TxFactory interface {
 	DeployContract(privKey cryptotypes.PrivKey, txArgs evmtypes.EvmTxArgs, deploymentData ContractDeploymentData) (common.Address, error)
 	// ExecuteContractCall executes a contract call with the provided private key
 	ExecuteContractCall(privKey cryptotypes.PrivKey, txArgs evmtypes.EvmTxArgs, callArgs CallArgs) (abcitypes.ResponseDeliverTx, error)
+	// GenerateSignedEthTx generates an Ethereum tx with the provided private key and txArgs but does not broadcast it.
+	GenerateSignedEthTx(privKey cryptotypes.PrivKey, txArgs evmtypes.EvmTxArgs) (signing.Tx, error)
 	// ExecuteEthTx builds, signs and broadcasts an Ethereum tx with the provided private key and txArgs.
 	// If the txArgs are not provided, they will be populated with default values or gas estimations.
 	ExecuteEthTx(privKey cryptotypes.PrivKey, txArgs evmtypes.EvmTxArgs) (abcitypes.ResponseDeliverTx, error)
@@ -75,6 +78,21 @@ func New(
 		network:              network,
 		ec:                   &ec,
 	}
+}
+
+// GenerateEthTx generates an Ethereum tx with the provided private key and txArgs but does not broadcast it.
+func (tf *IntegrationTxFactory) GenerateSignedEthTx(privKey cryptotypes.PrivKey, txArgs evmtypes.EvmTxArgs) (signing.Tx, error) {
+	msgEthereumTx, err := tf.createMsgEthereumTx(privKey, txArgs)
+	if err != nil {
+		return nil, errorsmod.Wrap(err, "failed to create ethereum tx")
+	}
+
+	signedMsg, err := signMsgEthereumTx(msgEthereumTx, privKey, tf.network.GetChainID())
+	if err != nil {
+		return nil, errorsmod.Wrap(err, "failed to sign ethereum tx")
+	}
+
+	return tf.buildSignedTx(signedMsg)
 }
 
 // CallContractAndCheckLogs is a helper function to call a contract and check the logs using
@@ -175,19 +193,14 @@ func (tf *IntegrationTxFactory) ExecuteEthTx(
 	priv cryptotypes.PrivKey,
 	txArgs evmtypes.EvmTxArgs,
 ) (abcitypes.ResponseDeliverTx, error) {
-	msgEthereumTx, err := tf.createMsgEthereumTx(priv, txArgs)
+	signedMsg, err := tf.GenerateSignedEthTx(priv, txArgs)
 	if err != nil {
-		return abcitypes.ResponseDeliverTx{}, errorsmod.Wrap(err, "failed to create ethereum tx")
+		return abcitypes.ResponseDeliverTx{}, errorsmod.Wrap(err, "failed to generate signed ethereum tx")
 	}
 
-	signedMsg, err := signMsgEthereumTx(msgEthereumTx, priv, tf.network.GetChainID())
+	txBytes, err := tf.encodeTx(signedMsg)
 	if err != nil {
-		return abcitypes.ResponseDeliverTx{}, errorsmod.Wrap(err, "failed to sign ethereum tx")
-	}
-
-	txBytes, err := tf.buildAndEncodeEthTx(signedMsg)
-	if err != nil {
-		return abcitypes.ResponseDeliverTx{}, errorsmod.Wrap(err, "failed to build and encode ethereum tx")
+		return abcitypes.ResponseDeliverTx{}, errorsmod.Wrap(err, "failed to encode ethereum tx")
 	}
 
 	res, err := tf.network.BroadcastTxSync(txBytes)
@@ -288,19 +301,19 @@ func (tf *IntegrationTxFactory) populateEvmTxArgs(
 	return txArgs, nil
 }
 
-func (tf *IntegrationTxFactory) buildAndEncodeEthTx(msg evmtypes.MsgEthereumTx) ([]byte, error) {
+func (tf *IntegrationTxFactory) encodeTx(tx sdktypes.Tx) ([]byte, error) {
 	txConfig := tf.ec.TxConfig
-	txBuilder := txConfig.NewTxBuilder()
-	signingTx, err := msg.BuildTx(txBuilder, tf.network.GetDenom())
-	if err != nil {
-		return nil, errorsmod.Wrap(err, "failed to build tx")
-	}
-
-	txBytes, err := txConfig.TxEncoder()(signingTx)
+	txBytes, err := txConfig.TxEncoder()(tx)
 	if err != nil {
 		return nil, errorsmod.Wrap(err, "failed to encode tx")
 	}
 	return txBytes, nil
+}
+
+func (tf *IntegrationTxFactory) buildSignedTx(msg evmtypes.MsgEthereumTx) (signing.Tx, error) {
+	txConfig := tf.ec.TxConfig
+	txBuilder := txConfig.NewTxBuilder()
+	return msg.BuildTx(txBuilder, tf.network.GetDenom())
 }
 
 // checkEthTxResponse checks if the response is valid and returns the MsgEthereumTxResponse